switch to using a central state context
