+++ src/components/TranscriptEditor/components/PreviewSection.tsx
@@ -1,0 +1,44 @@
+"use client";
+
+import { useSettings } from "@contexts/AppContext";
+import { ProjectTranscript } from "@types";
+import { generateSrt } from "@utils/export";
+import clsx from "clsx";
+import { useMemo } from "react";
+
+export function PreviewSection() {
+  const { projectTranscript, settings } = useSettings();
+
+  const srt = useMemo(() => {
+    if (!projectTranscript) return;
+    return generateSrt(projectTranscript, settings);
+  }, [projectTranscript, settings]);
+
+  return (
+    <div className="col-span-3 h-full rounded-lg bg-white p-4 shadow-md">
+      <h2 className="mb-4 text-xl font-bold select-none">Preview</h2>
+      <div className="flex h-[90cqh] flex-grow flex-col overflow-x-hidden overflow-y-scroll">
+        {srt?.subtitles.map((subtitle, index) => (
+          <div
+            key={index}
+            className="relative mb-4 overflow-clip rounded-md border-l-4 border-purple-500 bg-purple-50 p-3"
+          >
+            <div className="mb-1 text-xs font-semibold text-purple-500">Subtitle {index + 1}</div>
+            <div className="mb-2 text-xs text-gray-500">
+              {subtitle.startTime} → {subtitle.endTime}
+            </div>
+            <div className="relative w-fit text-base/6 whitespace-pre-line">
+              {settings.rule > 0 ? (
+                <div
+                  className="absolute h-full w-px bg-purple-600/30"
+                  style={{ left: `${settings.rule * 5}px` }}
+                />
+              ) : null}
+              <div className={clsx(settings.centerText && "text-center")}>{subtitle.text}</div>
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
